Moved all the stuff used or invoked by AggregatedTriggerWindow into AggregatedTriggerWindow.py
